Did more file refactoring, moved unittests inside sample folder.
