--- tests/Ad Hoc Testing/htmlTest.txt
@@ -1,7 +1,0 @@
-  <div style="font-size:10px">
-   <p style="font-family:Arial-BoldMT">
-   THIS IS BOLD TEXT
-   
- <div style="font-size:10px">
-   <p style="font-family:Arial"> 
-   THIS IS TEXT
