--- tests/test_tag_tree.py
@@ -1,18 +1,0 @@
-import unittest
-
-
-class TestTagTree(unittest.TestCase):
-    def setUp(self) -> None:
-        return super().setUp()
-
-    def tearDown(self) -> None:
-        return super().tearDown()
-    
-    def test_default1(self):
-        self.assertTrue(True)
-    
-    def test_default2(self):
-        self.assertTrue(True)
-
-if __name__ == '__main__':
-    unittest.main()
